Encapsulate moving settings for BloomPack
